improving architectural style
