+++ Java/src/main/java/com/gildedrose/application/ItemAdapter.java
@@ -1,4 +1,7 @@
-package com.gildedrose.domain.item;
+package com.gildedrose.application;
+
+import com.gildedrose.domain.Item;
+import com.gildedrose.domain.ItemType;
 
 public class ItemAdapter {
     private final Item item;
@@ -23,4 +26,9 @@ public class ItemAdapter {
     }
 
     public String toStringFull() { return this.toString() + ", " + itemType.getName(); }
+
+    public static ItemAdapter createItemAdapter(Item item) {
+        ItemType itemType = ItemType.findByValue(item.name);
+        return new ItemAdapter(itemType, item);
+    }
 }
