--- Java/src/main/java/com/gildedrose/domain/item/ItemAdapterResolver.java
@@ -1,9 +1,0 @@
-package com.gildedrose.domain.item;
-
-public final class ItemAdapterResolver {
-
-    public static ItemAdapter createItemAdapter(Item item) {
-        ItemType itemType = ItemType.findByValue(item.name);
-         return new ItemAdapter(itemType, item);
-    }
-}
