+++ Java/src/test/java/com/gildedrose/application/ItemAdapterTest.java
@@ -1,16 +1,17 @@
-package com.gildedrose.domain.item;
+package com.gildedrose.application;
 
+import com.gildedrose.domain.Item;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class ItemAdapterResolverTest {
+public class ItemAdapterTest {
 
     @Test
     void shouldCreateAgedBrieItemAdapter(){
         //given
         Item agedBrie = new Item("Aged Brie", 2, 0);
         //when
-        ItemAdapter result = ItemAdapterResolver.createItemAdapter(agedBrie);
+        ItemAdapter result = ItemAdapter.createItemAdapter(agedBrie);
         //then
         assertEquals(agedBrie.name, result.getItem().name);
         assertEquals(agedBrie.sellIn, result.getItem().sellIn);
@@ -23,7 +24,7 @@ public class ItemAdapterResolverTest {
         //given
         Item backstagePasses = new Item("Backstage passes to a TAFKAL80ETC concert", 15, 20);
         //when
-        ItemAdapter result = ItemAdapterResolver.createItemAdapter(backstagePasses);
+        ItemAdapter result = ItemAdapter.createItemAdapter(backstagePasses);
         //then
         assertEquals(backstagePasses.name, result.getItem().name);
         assertEquals(backstagePasses.sellIn, result.getItem().sellIn);
@@ -36,7 +37,7 @@ public class ItemAdapterResolverTest {
         //given
         Item sulfuras = new Item("Sulfuras, Hand of Ragnaros", 0, 80);
         //when
-        ItemAdapter result = ItemAdapterResolver.createItemAdapter(sulfuras);
+        ItemAdapter result = ItemAdapter.createItemAdapter(sulfuras);
         //then
         assertEquals(sulfuras.name, result.getItem().name);
         assertEquals(sulfuras.sellIn, result.getItem().sellIn);
@@ -49,7 +50,7 @@ public class ItemAdapterResolverTest {
         //given
         Item standardItem = new Item("Elixir of the Mongoose", 5, 7);
         //when
-        ItemAdapter result = ItemAdapterResolver.createItemAdapter(standardItem);
+        ItemAdapter result = ItemAdapter.createItemAdapter(standardItem);
         //then
         assertEquals(standardItem.name, result.getItem().name);
         assertEquals(standardItem.sellIn, result.getItem().sellIn);
@@ -62,7 +63,7 @@ public class ItemAdapterResolverTest {
         //given
         Item conjured = new Item("Conjured Mana Cake", 3, 6);
         //when
-        ItemAdapter result = ItemAdapterResolver.createItemAdapter(conjured);
+        ItemAdapter result = ItemAdapter.createItemAdapter(conjured);
         //then
         assertEquals(conjured.name, result.getItem().name);
         assertEquals(conjured.sellIn, result.getItem().sellIn);
