+++ Java/src/test/java/com/gildedrose/application/rules/AgedBrieRuleTest.java
@@ -1,9 +1,8 @@
-package com.gildedrose.application.aggedbrie;
+package com.gildedrose.application.rules;
 
-import com.gildedrose.application.agedbrie.AgedBrieRule;
-import com.gildedrose.domain.item.Item;
-import com.gildedrose.domain.item.ItemAdapter;
-import com.gildedrose.domain.item.ItemType;
+import com.gildedrose.domain.Item;
+import com.gildedrose.application.ItemAdapter;
+import com.gildedrose.domain.ItemType;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
